+++ R/build_teaching_dfs.R
@@ -1,4 +1,4 @@
-build_teaching_dfs <- function(
+# build_teaching_dfs <- function()
 
 d_df <-
   data.frame(
@@ -77,14 +77,16 @@ d_df <-
     )
   )
 
-)
 
+library(tidyverse)
+
+cols_to_drop <- c("X")
 
 d_df2 <-
     d_df |>
     mutate(Package = stringr::str_remove(source,
                                          "https://vincentarelbundock.github.io/Rdatasets/csv/") |>
              str_remove("/\\w+\\.csv$")) |>
-    select(-X)
+    select(any_of(cols_to_drop))
 
-write_csv(d_df2, "data/teaching_df.csv")
+#write_csv(d_df2, "data/teaching_df.csv")