+++ R/data-teaching_data.R
@@ -1,0 +1,9 @@
+#' Datasets for teaching
+#'
+#' @format A dataframe with `r nrow(teaching_data)` rows
+#'   and `r ncol(teaching_data)` cols
+#'
+#' @examples
+#' teaching_df
+#'
+"teaching_data"